Re-direct users to home page if accessing an invalid org or project
