+++ studio/components/interfaces/App/index.ts
@@ -1,0 +1,5 @@
+import PortalToast from './PortalToast'
+import GoTrueWrapper from './GoTrueWrapper/GoTrueWrapper'
+import RouteValidationWrapper from './RouteValidationWrapper'
+
+export { PortalToast, GoTrueWrapper, RouteValidationWrapper }
